server/requirements.txt
py_healthcheck==1.10.1
connexion==2.7.0
numpy==1.18.4
healthcheck==1.3.3
scikit-image==0.17.2
Flask==1.1.2
numpy==1.18.4
Keras==2.3.1
Pillow==7.1.2
kafka-python==2.0.1
python-dotenv==0.14.0